clean up pg types
